refactor: rename module to `_tabular_dataset`
