+++ tests/safeds/data/tabular/containers/_table/_tabular_dataset/test_from_table.py
