+++ tests/safeds/data/tabular/containers/_table/_tabular_dataset/test_sizeof.py
